requirements.txt
# Core web framework
Flask==2.3.3
Werkzeug==2.3.7

# Machine Learning and AI
tensorflow==2.13.0
transformers==4.33.2
torch==2.0.1
torchvision==0.15.2

# Image processing
Pillow==10.0.0
opencv-python==4.8.0.76
scikit-image==0.21.0

# Data science and numerical computing
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Database
sqlite3

# HTTP requests
requests==2.31.0

# Environment management
python-dotenv==1.0.0

# File handling
secure-filename==0.1.0

# Logging and monitoring
logging

# Development tools
pytest==7.4.2
pytest-flask==1.2.0

# Optional: For advanced NLP features
spacy==3.6.1
nltk==3.8.1

# Optional: For model deployment
gunicorn==21.2.0

# Image utilities
imageio==2.31.1

# JSON handling
jsonschema==4.19.0

# Date/time utilities
python-dateutil==2.8.2

# Mathematical operations
scipy==1.11.2

# Progress bars
tqdm==4.65.0

# Configuration management
configparser==5.3.0

# File system utilities
pathlib2==2.3.7

# Regular expressions (built-in)
re

# Operating system interface (built-in)
os

# Input/output operations (built-in)
io

# Base64 encoding (built-in)
base64

# Pickle serialization (built-in)
pickle

# Date and time (built-in)
datetime

# Secure random number generation
secrets

# UUID generation
uuid

# Threading
threading

# Collections
collections
